fix: remove circularity in imports between .wkrm and .utils
