+++ arches_orm/graphql/resources.py
@@ -18,7 +18,7 @@ import graphene
 from graphene_file_upload.scalars import Upload
 
 from aiodataloader import DataLoader
-from arches_orm.utils import attempt_well_known_resource_model, get_well_known_resource_model_by_class_name
+from arches_orm.wkrm import attempt_well_known_resource_model, get_well_known_resource_model_by_class_name
 from arches_orm.wkrm import WELL_KNOWN_RESOURCE_MODELS
 from arches_orm.wkrm import get_resource_models_for_adapter
